Reduced heap allocation used to compute the STFT of an input file.
This commit reduces the amount of copying used to compute the STFT of an
input file. Previously, various data was copied on the heap to perform
the computations naively. With this commit, we do things more in-place
using more complicated offset / length information, or at the very least
we do things on the stack to reduce malloc/free overhead.

This commit also adds some additional profiling code, when building in
debug mode, so the speed of our FFT functions can be analyzed and
improved.