Greatly reduced the amount of copying done by our FFT algorithm.
Our FFT algorithm doesn't quite compute things in-place, but it was
previously copying the ENTIRE dft list for each recursion on it. This
was highly inefficient. Updating this code speeds up our algorithm by a
factor of about 100, so now processing a 4 or 5 minute MP3 file takes
around 20 seconds when built with -O0.